Initial ident lexer

+++ crates/lexer/src/token.rs
@@ -6,5 +6,5 @@ use crate::TokenKind;
 pub struct Token<'a> {
     pub text: &'a str,
     pub kind: TokenKind,
-    pub span: Range<u32>,
+    pub span: Range<usize>,
 }
